Implemented authorization layer
This adds a basic authorization layer to Giftless. It allows for authentication adapters to return an identity object which can be authorized or not to do certain actions on certain organizations / repositories / objects. 

Currently, the batch API is the only place where authorization is checked (yet). 

The ‘allow_anon’ authenticator now provides two possible flavors: read_only and read_write.

WIP on adding JWT scopes based authorization and on adding additional authorization checks to other views.