Add basic mime type detection
Added new `get_mime_type()` method to `StreamingStorage` so that
different backends can implement it. For local storage we use the
builitin `mimetypes` module to guess the mime type from the filename,
for Azure we get it from the blob properties.

In the basic streaming transfer handler we read the value of this method
and add it to the `Content-type` header.