s3_external: force the upload Content-Type
git-lfs as of v2.5.0 [1] detects the file Content-Type and provides it
during PUT requests to the storage when using the basic transfer. If we
don't include the same `Content-Type` value in the S3 signature
generation, then S3 rejects it during upload.

Unfortunately git-lfs doesn't pass the `Content-Type` value to the Batch
request, so it can't be included in S3 signature generation. And while
setting `lfs.contenttype=false` in the client config forces sending
`application/octet-stream` [2], S3 isn't happy with that either.

If the Batch response upload action includes a `Content-Type` header
then git-lfs will always use that for the PUT request, so always return
`application/octet-stream`, and also use it for signature generation.

[1] https://github.com/git-lfs/git-lfs/pull/3137
[2] https://github.com/git-lfs/git-lfs/pull/3163