chore(github): make user auth caching more robust + tests
Setting any threshold for the lowest value for the auth cache size might still not guarantee (in some rough timing conditions), the entry won't disappear before it's read for the first time by the core logic.
The change adds a short "hot/proxy" cache for new entries, that are not yet ever read by the core logic via is_authorized(). When that's done, only then the entry is moved to the regular TTL/LRU cache.