fix(preauth): Improve action's 'authenticated' flag
The LFS batch response always included authenticated = true, even when the default "PRE_AUTHORIZED_ACTION_PROVIDER" was disabled (= null). This led the git client to forever retry the download/upload attempts without realizing it should try the credentials from its creds store.

This comes in handy with the GitHub auth provider which can handle both the usual authentication and the same for the streaming transfer adapter.